requirements.txt
Django==2.1.14 #This requirement must remain pinned at 2.1.14 until https://github.com/PyMySQL/PyMySQL/issues/790 is resolved
PyMySQL==0.9.3
mysqlclient==1.4.6
django-markdownx==3.0.1
django-stdimage==5.0.2
pillow==7.0.0
django-mathjax==0.0.8
django-storages[google]==1.9.1
django-summernote==0.8.11.6